% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_distributions.R
\name{plot_distributions}
\alias{plot_distributions}
\title{Plot distributions of data columns}
\usage{
plot_distributions(data, column = NULL)
}
\arguments{
\item{data}{A data frame}

\item{column}{Optional specific column name}
}
\value{
ggplot2 visualizations
}
\description{
Plot distributions of data columns
}
\examples{
plot_distributions(iris)
}
