% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_data.R
\name{report_data}
\alias{report_data}
\title{Generate comprehensive data report}
\usage{
report_data(data)
}
\arguments{
\item{data}{A data frame}
}
\value{
A list with comprehensive data summary
}
\description{
Generate comprehensive data report
}
\examples{
report_data(iris)
}
