% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suggest_approach.R
\name{suggest_approach}
\alias{suggest_approach}
\title{Suggest modeling approach based on data characteristics}
\usage{
suggest_approach(data_report)
}
\arguments{
\item{data_report}{Output from report_data function or a data frame}
}
\value{
Text suggestions for modeling approach
}
\description{
Suggest modeling approach based on data characteristics
}
\examples{
report <- report_data(iris)
suggest_approach(report)
}
